refactor: extract metadata cache and flush kernel into dev/cache.rs
Move the IO-op wrappers (call_*), table loaders, L2-slice cache
management and the soft-update flush logic (flush_*, flush_meta) into a
new dev/cache.rs submodule as an impl extension of Qcow2Dev. The kernel
methods invoked by the read/write/alloc/check/discard submodules are
widened to pub(crate); call_read keeps its #[inline] and mark_need_flush
keeps #[inline(always)]. dev/mod.rs is now just the struct, constructor,
lifecycle helpers and prep/flush flags. No logic changes.
